extract configure command to its own file
