+++ src/fake-aws-creds.json
@@ -1,5 +1,5 @@
 {
-  "accessKey": "s",
-  "secretAccessKey": "d",
-  "cloudfrontARN": "y"
+  "accessKey": "alex",
+  "secretAccessKey": "rivie",
+  "cloudfrontARN": "we outchea"
 }
